adapt to openssl connection
